feat: 가격 API 수정에 다른 request, response 수정
