+++ iOS/airbnb/Network/TravelNetworkDispatcher.swift
@@ -18,22 +18,3 @@ class TravelNetworkDispatcher: NetworkDispatcher {
     }
 
 }
-
-class PriceNetworkDispatcher: NetworkDispatcher {
-    
-    func execute<T>(decodeType: T.Type) -> AnyPublisher<T, NetworkError> where T : Decodable {
-        
-        guard let url = EndPoint.search.url else {
-            return Fail(error: NetworkError.invalidURL).eraseToAnyPublisher()
-        }
-        var componets = URLComponents(url: url, resolvingAgainstBaseURL: false)
-        componets?.path.append("/prices/")
-        
-        let local = URLQueryItem(name: "locationName", value: "서울")
-        componets?.queryItems = [local]
-        guard let componentURL = componets?.url else {
-            return Fail(error: NetworkError.invalidURL).eraseToAnyPublisher()
-        }
-        return request(url : componentURL)
-    }
-}
